+++ tests/testthat/test-add_hierarchical_count_row.R
@@ -1,0 +1,17 @@
+test_that("add_hierarchical_count_row() works", {
+  expect_snapshot(
+    cards::ADAE |>
+      # subset the data for a shorter example table
+      dplyr::slice(1:10) |>
+      tbl_hierarchical(
+        by = "TRTA",
+        variables = AEDECOD,
+        denominator = cards::ADSL |> dplyr::rename(TRTA = TRT01A),
+        id = "USUBJID",
+        overall_row = TRUE
+      ) |>
+      add_hierarchical_count_row(.after = 1L) |>
+      as.data.frame(col_label = FALSE) |>
+      dplyr::slice(2)
+  )
+})