+++ tests/testthat/test-tbl_hierarchical_rate_and_count.R
@@ -5,16 +5,15 @@ ADAE_subset <- cards::ADAE |>
   dplyr::filter(AEBODSYS %in% unique(AEBODSYS)[1:2]) |>
   dplyr::slice(1:10)
 
-test_that("tbl_ae_rate_and_count() works", {
+test_that("tbl_hierarchical_rate_and_count() works", {
   withr::local_options(list(width = 220))
   expect_silent(
     tbl <-
       ADAE_subset |>
-      tbl_ae_rate_and_count(
+      tbl_hierarchical_rate_and_count(
         denominator = cards::ADSL |> dplyr::rename(TRTA = TRT01A),
         by = TRTA,
-        ae = AEDECOD,
-        soc = AEBODSYS,
+        variables = c(AEBODSYS, AEDECOD)
       ) |>
       add_overall(last = TRUE)
   )
@@ -30,65 +29,30 @@ test_that("tbl_ae_rate_and_count() works", {
 
   # snapshot of the table
   expect_snapshot(as.data.frame(tbl))
-})
 
-test_that("tbl_ae_rate_and_count(hlt)", {
-  withr::local_options(list(width = 220))
-
+  # testing with a high level term
   expect_silent(
     tbl <-
       ADAE_subset |>
-      tbl_ae_rate_and_count(
+      tbl_hierarchical_rate_and_count(
         denominator = cards::ADSL |> dplyr::rename(TRTA = TRT01A),
         by = TRTA,
-        ae = AEDECOD,
-        soc = AEBODSYS,
-        hlt = AEHLT
+        variables = c(AEBODSYS, AEHLT, AEDECOD)
       ) |>
       add_overall(last = TRUE)
   )
-
   # snapshot of the table
   expect_snapshot(as.data.frame(tbl))
 })
 
-test_that("tbl_ae_rate_and_count(filter)", {
+test_that("tbl_hierarchical_rate_and_count(sort)", {
   expect_silent(
     tbl <-
       ADAE_subset |>
-      tbl_ae_rate_and_count(
+      tbl_hierarchical_rate_and_count(
         denominator = cards::ADSL |> dplyr::rename(TRTA = TRT01A),
         by = TRTA,
-        ae = AEDECOD,
-        soc = AEBODSYS,
-        filter = sum(n) / sum(N) > 0.01
-      ) |>
-      add_overall(last = TRUE)
-  )
-
-  # check that all rows have a percent larger than 1%
-  expect_true(
-    tbl$table_body |>
-      dplyr::filter(str_detect(stat_0, "%") & variable == "AEDECOD") |>
-      dplyr::pull(stat_0) |>
-      str_extract("(?<=\\().*?(?=\\))") |> # extract the percent between the parantheses
-      str_remove("%") %>%
-      {
-        as.numeric(.) > 1.0
-      } |> # styler off
-      all()
-  )
-})
-
-test_that("tbl_ae_rate_and_count(sort)", {
-  expect_silent(
-    tbl <-
-      ADAE_subset |>
-      tbl_ae_rate_and_count(
-        denominator = cards::ADSL |> dplyr::rename(TRTA = TRT01A),
-        by = TRTA,
-        ae = AEDECOD,
-        soc = AEBODSYS,
+        variables = c(AEBODSYS, AEDECOD),
         sort = "descending"
       ) |>
       add_overall(last = TRUE)